Added code for voltage-angle adjustment

+++ app/src/main/java/de/kai_morich/simple_usb_terminal/SensorHelper.java
@@ -60,18 +60,8 @@ public class SensorHelper extends Service implements SensorEventListener {
         return START_STICKY;
     }
 
-    public static void setHeading(float pot_angle)  {
-        //small alpha filter
-        //double difference = Math.abs(heading - pot_angle);
-        //if(heading <= 0) {
-        //    heading = pot_angle;
-        //}
-        //else if(difference < 50 ) {
-        //    heading = pot_angle;
-        //}
+    public static void setHeading(double pot_angle)  {
         heading = pot_angle;
-
-
     }
 
     /**
